requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.1.1
gunicorn==19.6.0

# Static and Media Storage
# ------------------------------------------------
boto==2.41.0
django-storages-redux==1.3.2
Collectfast==0.2.3

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.4.2




